extract a common file, and add a S3 upload method
